requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
asyncpg==0.30.0
    # via -r requirements.in
click==8.1.8
    # via uvicorn
fastapi==0.115.8
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.38
    # via -r requirements.in
starlette==0.45.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.34.0
    # via -r requirements.in
